Add result caching into TBA API
This reduces the number of requests the user makes to the TBA API by
storing any request to a specific resource and serving the result to
future requests.

Currently this cache only works on a per-session basis. Creating a
file/browser cache would create a better situation. However, the cache
also needs to be able to be cleared.